+++ internal/storage/metricstore/elasticsearch/query_builder.go
@@ -10,6 +10,7 @@ import (
 	"time"
 
 	"github.com/olivere/elastic/v7"
+	"go.uber.org/zap"
 
 	"github.com/jaegertracing/jaeger-idl/model/v1"
 	es "github.com/jaegertracing/jaeger/internal/storage/elasticsearch"
@@ -31,15 +32,20 @@ const (
 // QueryBuilder is responsible for constructing Elasticsearch queries (bool and aggregation)
 // based on provided parameters and executing them to retrieve raw search results.
 type QueryBuilder struct {
-	client es.Client
-	cfg    config.Configuration
+	client           es.Client
+	cfg              config.Configuration
+	timeRangeIndices spanstore.TimeRangeIndexFn
 }
 
 // NewQueryBuilder creates a new QueryBuilder instance.
-func NewQueryBuilder(client es.Client, cfg config.Configuration) *QueryBuilder {
+func NewQueryBuilder(client es.Client, cfg config.Configuration, logger *zap.Logger) *QueryBuilder {
 	return &QueryBuilder{
 		client: client,
 		cfg:    cfg,
+		timeRangeIndices: spanstore.LoggingTimeRangeIndexFn(
+			logger,
+			spanstore.TimeRangeIndicesFn(cfg.UseReadWriteAliases, cfg.ReadAliasSuffix, cfg.RemoteReadClusters),
+		),
 	}
 }
 
@@ -110,8 +116,7 @@ func (*QueryBuilder) buildTimeSeriesAggQuery(params metricstore.BaseQueryParamet
 // Execute runs the Elasticsearch search with the provided bool and aggregation queries.
 func (q *QueryBuilder) Execute(ctx context.Context, boolQuery elastic.BoolQuery, aggQuery elastic.Aggregation, timeRange TimeRange) (*elastic.SearchResult, error) {
 	indexName := q.cfg.Indices.IndexPrefix.Apply("jaeger-span-")
-	timeRangeIndicesFn := spanstore.TimeRangeIndicesFn(q.cfg.UseReadWriteAliases, q.cfg.ReadAliasSuffix, q.cfg.RemoteReadClusters)
-	indices := timeRangeIndicesFn(
+	indices := q.timeRangeIndices(
 		indexName,
 		q.cfg.Indices.Services.DateLayout,
 		time.UnixMilli(timeRange.extendedStartTimeMillis).UTC(),
