+++ internal/storage/v1/elasticsearch/spanstore/index_utils.go
@@ -7,70 +7,8 @@ import (
 	"time"
 )
 
-type TimeRangeIndexFn func(indexName string, indexDateLayout string, startTime time.Time, endTime time.Time, reduceDuration time.Duration) []string
-
-func TimeRangeIndicesFn(useReadWriteAliases bool, readAliasSuffix string, remoteReadClusters []string) TimeRangeIndexFn {
-	suffix := ""
-	if useReadWriteAliases {
-		if readAliasSuffix != "" {
-			suffix = readAliasSuffix
-		} else {
-			suffix = "read"
-		}
-	}
-	return addRemoteReadClusters(
-		getTimeRangeIndexFn(useReadWriteAliases, suffix),
-		remoteReadClusters,
-	)
-}
-
 // returns index name with date
 func indexWithDate(indexPrefix, indexDateLayout string, date time.Time) string {
 	spanDate := date.UTC().Format(indexDateLayout)
 	return indexPrefix + spanDate
-}
-
-func getTimeRangeIndexFn(useReadWriteAliases bool, readAlias string) TimeRangeIndexFn {
-	if useReadWriteAliases {
-		return func(indexPrefix, _ /* indexDateLayout */ string, _ /* startTime */ time.Time, _ /* endTime */ time.Time, _ /* reduceDuration */ time.Duration) []string {
-			return []string{indexPrefix + readAlias}
-		}
-	}
-	return timeRangeIndices
-}
-
-// Add a remote cluster prefix for each cluster and for each index and add it to the list of original indices.
-// Elasticsearch cross cluster api example GET /twitter,cluster_one:twitter,cluster_two:twitter/_search.
-func addRemoteReadClusters(fn TimeRangeIndexFn, remoteReadClusters []string) TimeRangeIndexFn {
-	return func(indexPrefix string, indexDateLayout string, startTime time.Time, endTime time.Time, reduceDuration time.Duration) []string {
-		jaegerIndices := fn(indexPrefix, indexDateLayout, startTime, endTime, reduceDuration)
-		if len(remoteReadClusters) == 0 {
-			return jaegerIndices
-		}
-
-		for _, jaegerIndex := range jaegerIndices {
-			for _, remoteCluster := range remoteReadClusters {
-				remoteIndex := remoteCluster + ":" + jaegerIndex
-				jaegerIndices = append(jaegerIndices, remoteIndex)
-			}
-		}
-
-		return jaegerIndices
-	}
-}
-
-// timeRangeIndices returns the array of indices that we need to query, based on query params
-func timeRangeIndices(indexName, indexDateLayout string, startTime time.Time, endTime time.Time, reduceDuration time.Duration) []string {
-	var indices []string
-	firstIndex := indexWithDate(indexName, indexDateLayout, startTime)
-	currentIndex := indexWithDate(indexName, indexDateLayout, endTime)
-	for currentIndex != firstIndex && endTime.After(startTime) {
-		if len(indices) == 0 || indices[len(indices)-1] != currentIndex {
-			indices = append(indices, currentIndex)
-		}
-		endTime = endTime.Add(reduceDuration)
-		currentIndex = indexWithDate(indexName, indexDateLayout, endTime)
-	}
-	indices = append(indices, firstIndex)
-	return indices
 }
