+++ internal/storage/v1/elasticsearch/spanstore/reader.go
@@ -144,7 +144,7 @@ func NewSpanReader(p SpanReaderParams) *SpanReader {
 		serviceIndexPrefix:      p.IndexPrefix.Apply(serviceIndexBaseName),
 		spanIndex:               p.SpanIndex,
 		serviceIndex:            p.ServiceIndex,
-		timeRangeIndices: getLoggingTimeRangeIndexFn(
+		timeRangeIndices: LoggingTimeRangeIndexFn(
 			p.Logger,
 			TimeRangeIndicesFn(p.UseReadWriteAliases, p.ReadAliasSuffix, p.RemoteReadClusters),
 		),
@@ -157,9 +157,11 @@ func NewSpanReader(p SpanReaderParams) *SpanReader {
 	}
 }
 
+type TimeRangeIndexFn func(indexName string, indexDateLayout string, startTime time.Time, endTime time.Time, reduceDuration time.Duration) []string
+
 type sourceFn func(query elastic.Query, nextTime uint64) *elastic.SearchSource
 
-func getLoggingTimeRangeIndexFn(logger *zap.Logger, fn TimeRangeIndexFn) TimeRangeIndexFn {
+func LoggingTimeRangeIndexFn(logger *zap.Logger, fn TimeRangeIndexFn) TimeRangeIndexFn {
 	if !logger.Core().Enabled(zap.DebugLevel) {
 		return fn
 	}
@@ -167,6 +169,50 @@ func getLoggingTimeRangeIndexFn(logger *zap.Logger, fn TimeRangeIndexFn) TimeRan
 		indices := fn(indexName, indexDateLayout, startTime, endTime, reduceDuration)
 		logger.Debug("Reading from ES indices", zap.Strings("index", indices))
 		return indices
+	}
+}
+
+func TimeRangeIndicesFn(useReadWriteAliases bool, readAliasSuffix string, remoteReadClusters []string) TimeRangeIndexFn {
+	suffix := ""
+	if useReadWriteAliases {
+		if readAliasSuffix != "" {
+			suffix = readAliasSuffix
+		} else {
+			suffix = "read"
+		}
+	}
+	return addRemoteReadClusters(
+		getTimeRangeIndexFn(useReadWriteAliases, suffix),
+		remoteReadClusters,
+	)
+}
+
+func getTimeRangeIndexFn(useReadWriteAliases bool, readAlias string) TimeRangeIndexFn {
+	if useReadWriteAliases {
+		return func(indexPrefix, _ /* indexDateLayout */ string, _ /* startTime */ time.Time, _ /* endTime */ time.Time, _ /* reduceDuration */ time.Duration) []string {
+			return []string{indexPrefix + readAlias}
+		}
+	}
+	return timeRangeIndices
+}
+
+// Add a remote cluster prefix for each cluster and for each index and add it to the list of original indices.
+// Elasticsearch cross cluster api example GET /twitter,cluster_one:twitter,cluster_two:twitter/_search.
+func addRemoteReadClusters(fn TimeRangeIndexFn, remoteReadClusters []string) TimeRangeIndexFn {
+	return func(indexPrefix string, indexDateLayout string, startTime time.Time, endTime time.Time, reduceDuration time.Duration) []string {
+		jaegerIndices := fn(indexPrefix, indexDateLayout, startTime, endTime, reduceDuration)
+		if len(remoteReadClusters) == 0 {
+			return jaegerIndices
+		}
+
+		for _, jaegerIndex := range jaegerIndices {
+			for _, remoteCluster := range remoteReadClusters {
+				remoteIndex := remoteCluster + ":" + jaegerIndex
+				jaegerIndices = append(jaegerIndices, remoteIndex)
+			}
+		}
+
+		return jaegerIndices
 	}
 }
 
@@ -178,6 +224,22 @@ func getSourceFn(maxDocCount int) sourceFn {
 			Sort("startTime", true).
 			SearchAfter(nextTime)
 	}
+}
+
+// timeRangeIndices returns the array of indices that we need to query, based on query params
+func timeRangeIndices(indexName, indexDateLayout string, startTime time.Time, endTime time.Time, reduceDuration time.Duration) []string {
+	var indices []string
+	firstIndex := indexWithDate(indexName, indexDateLayout, startTime)
+	currentIndex := indexWithDate(indexName, indexDateLayout, endTime)
+	for currentIndex != firstIndex {
+		if len(indices) == 0 || indices[len(indices)-1] != currentIndex {
+			indices = append(indices, currentIndex)
+		}
+		endTime = endTime.Add(reduceDuration)
+		currentIndex = indexWithDate(indexName, indexDateLayout, endTime)
+	}
+	indices = append(indices, firstIndex)
+	return indices
 }
 
 // GetTraces takes a traceID and returns a Trace associated with that traceID
